Extract `collection/point_ops.rs`
